Fix quiclook not found for S2 L1C
There are two bugs in the way it looks for the quicklook in the S2 .SAFE scene:
1. the pattern was missing a wildcard
2. The TCI is 10m resolution for L1C, thus better use PVI which is the preview image (although 320m resolution)

I did not change regex pattern for L2A products, although the PVI image also exists. I would recommend using it also instead of TCI_60m as this last one can be quite heavy (several MB). But I let you make that change if you feel like it is the good solution.